gulp-concat integration, refactored oauth.js to self invoked func
