+++ gulpfile.js
@@ -3,7 +3,8 @@
 const gulp = require('gulp');
 
 require('./styles/gulp/tasks/style')(gulp);
+require('./styles/gulp/tasks/js')(gulp);
 require('./styles/gulp/tasks/watch')(gulp);
 
-gulp.task('default', ['style:lint','style:dev', 'watch']);
+gulp.task('default', ['style:lint','style:dev', 'js', 'watch']);
 gulp.task('release', ['style:release']);
