+++ package.json
@@ -21,6 +21,7 @@
   "devDependencies": {
     "gulp": "~3.9.1",
     "gulp-autoprefixer": "^3.1.1",
+    "gulp-concat": "^2.6.1",
     "gulp-connect": "~5.0.0",
     "gulp-rename": "^1.2.2",
     "gulp-sass": "~2.3.2",
