+++ styles/gulp/options/paths.js
@@ -6,4 +6,9 @@ module.exports = {
         src_watch: 'styles/sass/**/*.scss',
         dist: './'
     },
+
+    js: {
+        src: 'js/**/*.js',
+        dest: './'
+    }
 }
