+++ system/application/views/footer.php
@@ -1,8 +1,7 @@
         </div>
     </div>
     <?php require('footer_links.php') ?>
-    <script src="/comment.js" type="text/javascript"></script>
-    <script src="/oauth.js" type="text/javascript"></script>
+    <script src="/index.js" type="text/javascript"></script>
 </body>
 </html>
