+++ system/application/views/footer_login.php
@@ -1,5 +1,5 @@
     </div>
     <?php require 'footer_links.php' ?>
-    <script src="/oauth.js" type="text/javascript"></script>
+    <script src="/index.js" type="text/javascript"></script>
 </body>
 </html>
