+++ system/application/views/oauth/_login-button.php
@@ -1,17 +1,17 @@
 
-<a id="lnkLoginGoogle" href="#" class="cs-button cs-button-social cs-button-google">
+<a id="lnkLoginGoogle" href="#" class="cs-button cs-button-social cs-button-google oauth-link">
     <svg>
         <use xlink:href="/resources/icons/sprite.svg#cs-google">
     </svg>
     <span>Google</span>
 </a>
-<a id="lnkLoginFacebook" href="#" class="cs-button cs-button-social cs-button-facebook">
+<a id="lnkLoginFacebook" href="#" class="cs-button cs-button-social cs-button-facebook oauth-link">
     <svg>
         <use xlink:href="/resources/icons/sprite.svg#cs-facebook">
     </svg>
     <span>Facebook</span>
 </a>
-<a id="lnkLoginGitHub" href="#" class="cs-button cs-button-social cs-button-github">
+<a id="lnkLoginGitHub" href="#" class="cs-button cs-button-social cs-button-github oauth-link">
     <svg>
         <use xlink:href="/resources/icons/sprite.svg#cs-github">
     </svg>
